[ch69] Rename components
