+++ src/components/ProductContent/ProductMedia.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const ProductMedia = () => {
+  return (
+    <div className="product-media">
+      <div className="product-media__image-container">
+        <img src="#" alt="product-media__image" />
+      </div>
+    </div>
+  )
+}
+
+export default ProductMedia
