--- src/components/ProductView/product-content.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const ProductContent = () => {
-  return (
-    <div className="product-content">
-      <div className="product-content__image">
-        <img src="#" alt="product-image" />
-      </div>
-    </div>
-  )
-}
-
-export default ProductContent
