+++ src/styles/components/__components-dir.scss
@@ -7,7 +7,7 @@
 @import './sign-up-success-form';
 @import './form';
 @import './product-list';
-@import './product-view';
+@import './product-media';
 @import './product-content';
 @import './product-information';
 @import './product-detail';
