+++ src/styles/components/_product-media.scss
@@ -1,0 +1,7 @@
+.product-media {
+  &__image {
+    margin-left: auto;
+    margin-right: auto;
+    width: 40%;
+  }
+}
