--- src/styles/components/_product-view.scss
@@ -1,35 +1,0 @@
-.product-view {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: repeat(3, 1fr);
-  grid-template-areas:
-    'content'
-    'detail'
-    'information';
-
-  @include respond-above('lg') {
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(2, 1fr);
-    grid-template-areas:
-      'content detail'
-      'information information';
-  }
-
-  & > .product-view__content {
-    grid-area: content;
-
-    background-color: green;
-  }
-
-  & > .product-view__detail {
-    grid-area: detail;
-
-    background-color: royalblue;
-  }
-
-  & > .product-view__information {
-    grid-area: information;
-
-    background-color: powderblue;
-  }
-}
