+++ src/styles/layouts/__layouts-dir.scss
@@ -2,4 +2,4 @@
 @import './layout-authentication';
 @import './layout-about-us';
 @import './layout-products';
-@import './layout-product-detail';
+@import './layout-product-view';
