+++ src/styles/layouts/_layout-product-view.scss
@@ -1,4 +1,4 @@
-.layout-product-detail {
+.layout-product-view {
   display: grid;
   grid-template-columns: 1fr;
   grid-template-rows: rem(40px) 1fr rem(200px);
